Update patch set 1
Patch Set 1:

> > so it's not amazingly optimized) it would be a pretty big performance hit I'd expect.
> 
> This is a pure conjecture on your part.

It's not.

[ 26830] get_safemode_option("disable_user_addons", 0x712451c0, [0x20]) = 0x80000007 Name not found (54 us)

By comparison, creating a memory area takes *1/4* that time:

[ 26830] create_area("heap", 0xe0e220, 0x1, 0x40000, 0x0, 0x33) = 0x000f72f7 (13 us)

And of course get_safemode_option returns only the string, so we need to test it against all the possible strings, as is done already in the Storage Kit for instance (http://xref.plausible.coop/source/xref/haiku/src/kits/storage/AddOnMonitorHandler.cpp#130).

My testing shows the strcmps add another 100us in the best case. So then we would have another 125-150us being spent on a syscall + string comparison on every application startup. That's a lot of cycles to burn. This change, on the other hand, will add <= 1 us for every application startup.

Patch-set: 1