Update patch set 2
Patch Set 2:

> > The exact same thing is true of glibc right now. Why is musl any different here?
> 
> Sure, this is true for every libc which would want to replace our custom glibc/libc/libroot combination.

No, I mean there is no reason to -1 this on the basis that it will make things more difficult to maintain, because the exact situation you are worried about we already have with glibc. The libc needs to remain in tree, we really cannot outsource that; so this problem we are going to stay stuck with.

Patch-set: 2