Update patch set 2
Patch Set 2:

> No, I mean there is no reason to -1 this on the basis that it will make things more difficult to maintain, because the exact situation you are worried about we already have with glibc. The libc needs to remain in tree, we really cannot outsource that; so this problem we are going to stay stuck with.

I want to be able to go back to a working libroot when things go wrong because they weren't tested with the full range of use cases. This patch doesn't allow this.
As you say we already have a situation with glibc in our tree, doing the same and importing musl parts (without versioning!!) doesn't improve it.

Patch-set: 2