Update patch set 2
Patch Set 2:

You say both that you want to lower the maintenance, and that no one has touched these files for the last 10 years, because indeed, they don't need any maintenance. There aren't this many bugs you can find and fix in an strcpy implementation.

So I agree with Jérôme here, if we want to switch to a standard libc, that would be one packaged using haikuporter, built with its own buildsystem and unit tests. Otherwise, we are still using a fork of another C library, and we do not reduce any maintenance effort. All the testing, keeping update with upstream, and investigating bugs will still be on us.

Patch-set: 2