Update patch set 1
Patch Set 1:

Actually I just realized that refactor would not make a lot of sense for the fallback either, as this would be needlessly restrictive in the case where the environment variables were "cleaned out" for some reason. So I still think the current implementation is fine as-is.

Patch-set: 1
Label: Code-Review=0