Update patch set 1
Patch Set 1:

> arch_mmu was already changed to use addr_t for arm. In fact I did a quick search for mmu_get_virtual_mapping and all the other ones I could find use addr_t as well. Does that mean that the other instances should be phys_addr_t as well?

The types specify what kind of address the type holds. phys_addr_t holds, as the name implies, a pointer to physical memory, whereas addr_t holds a pointer to virtual memory (almost always that of the kernel or of an application currently syscalling into the kernel.)

There's also the difference that on systems with PAE, the phys_addr_t is 64-bit while addr_t is of course 32-bit.

> As for the symbol issues the main one I was having a problem with was std::remove. Any place that included cstdio wouldn't build. I did some digging and typically when that happens (among issues for other projects I reviewed) it was related to the --std flag. I modified the Jamfile to use the same std flag as the x86 Jamfile and that resolved the error.

Well, I can't really verify this either way, because currently ARM needs to be bootstrapped again following the GCC and ICU upgrades, which currently fails (https://dev.haiku-os.org/ticket/14842).

Patch-set: 1