sparc: fix openfirmware entry point.
According to the IEEE standard documentation, the OpenFirmware entry
point should be in %o3. But that doesn't work, and both FreeBSD and
NetBSD expect it in %o4 (5th argument of the function).

The bootloader now attempts to call into openfirmware, but we now have
to adjust the parameters for all the calls to use 64bit int in the
passed structures, to match the 64bit forth cell size on sparc.