Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Perhaps it makes more sense to move _start into the arch code?

I tried, but BootMergeObject does not see that _start is the entry point for everything, and it does not use --whole-archive on the .a files added to it, so _start is not included in the .o and then linking fails.

It would make more sense, if we can get it to work.

Patch-set: 1