Update patch set 2
Patch Set 2:

> Is there any program out there that actually use this? musl libc's version of this always returns 0 and there aren't any complaint about that so far that I know of.

Given the specs, the function can't be used for anything useful. Returning 0 complies with POSIX since "This volume of POSIX.1-2017 does not define the domain in which the return value is unique." So "unique for the current machine" is fine.

I could imagine using something like this eg. in a syslog system, where an admin has control on all the machines and can give them an host id. Since the work is done already, I don't see a problem in merging it.

Patch-set: 2