kernel: Reinstate the USER_ADDRESS check in ioctl.
Thinking over this carefully, I realized that adding checks to
every ioctl implementation in every driver would be very prohibitive,
because there, one has to check is_called_via_syscall() in addition
to IS_USER_ADDRESS(), and this would have to be done in every case.
So that would take a massive amount of work, and it would be
very easy to miss a case.

Thankfully, the "size" flag is Haiku-specific, i.e. it's not in POSIX.
So we can check that if size != 0 *and* the buffer is a user-address,
then return an error. This should suffice to prevent exploitation
by passing in kernel buffers.