sparc: boot cpu support
There is no /cpus in the device tree on my machine, the CPUs are instead
directly under the root. So search for them there.

There is also no timebase property, and according to the docs, the
RDTICK instruction just gets CPU clock ticks, so set them to the same
value for now.

The two CPUs in my reference machine are succesfully found now, yay!

Change-Id: I5fbb4cfc98ee3b68cb2aa810816e0054a56d52d0