Update patch set 1
Patch Set 1:

I realized another reason we must do the USER_ADDRESS check here and cannot do it in the drivers: because the drivers do not know directly, in is_called_via_syscall, if it's really the kernel that is passing a buffer into them legitimately, unless SyscallFlagUnsetter is used. So then every invocation of every driver ioctl() in *any* part of kernel code would have to use a SyscallFlagUnsetter. No, we really must just do the check inside the syscall.

If we can't do it by checking "size", then we have to do it by handling each of the potential "command"s that can have non-user buffers and skipping the check for them.

Patch-set: 1