Update patch set 1
Patch Set 1:

Well, now that we know what a fix is, I'd rather we spend the time to find the proper one rather than just merging this :)

I should have a good amount of time this weekend to investigate this. Now that I know what causes it, finding how to do this properly should not be too hard.

Patch-set: 1