Update patch set 1
Patch Set 1:

Yes, I forgot that the toolbar can be on any side of the window.

Using a separator seems like an overkill to me - it would require adding another GroupView inside just for this. Also, the idea is to have it look like BMenu border, and while separator looks the same at the moment, it does not have to going forward.

I will revise the patch to accomodate toolbars on different sides of the window, but I would prefer to avoid adding another layout inside.

Patch-set: 1