BFS: Eliminate invalid use of non-static data member
clang complains about this line, some explination:
https://stackoverflow.com/questions/29359661/clang-error-invalid-use-of-non-static-data-member
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2007/n2253.html

Referencing a non-static data member inside a struct was retroactively
made undefined in C++98 and only allowed on C++11 even though gcc
permits the code to compile. In this case it was an easy fix to get
the size from the instance instead.

No change to functionality intended.