Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Actually we should be able to do this in the Pipe destructor, I think. But the memory management in here is a little strange to say the least.
> 
> Doing it in the destructor will not work because there are virtual methods and the subclasses destructors will already have been called, clearing the overriden methods from the vtable: https://www.artima.com/cppsource/nevercall.html

Did you even look at the Pipe destructor? It already calls CancelQueuedTransfers by invoking the BusManager directly, which obviously is valid.

But reviewing the code more, this solution is obviously wrong. We have to use the Force parameter here, as the driver will already be detached. The correct solution is to fix usb_raw, see my comments o the ticket.

Patch-set: 1