Update patch set 1
Patch Set 1: Code-Review-2

There is no real performance or other benefit here besides maybe a small bit of readability. But there is a rather significant downside that the caches are not named for the classes they are caching now. There are potential workarounds for that, but I still think the existing solution is the best.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2