Update patch set 3
Patch Set 3: Code-Review+2

> Patch Set 3:
> 
> I have compared the sizes (on both archs), and they didn't change with this patchset.
> 
> A lot more space could be reclaimed by reordering the members, but I didn't do it because there is potential for subtle breakage (inlined functions would reference them by position, I think).

That would be right, but these fields are all private and there are no inline functions in this class (we avoid them in public apis because of this, in fact). So, it should be safe.

However, it's unrelated to this change and may have other ill effects: on cache performance, and whenusing bitfields, there is also a risk of problems with concurrent access to different bits from multiple threads in situations that are currently safe.

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2