XHCI: WIP attempt at isochronous transfers support.
Starting seems to be rather flaky; i.e. the first few buffers
hit odd timeouts, but after that it seems to work OK, at least
in QEMU. Likely the remaining bugs are in frame calculation,
which I am likely doing very incorrectly; but in my defense
all extant drivers and the specification are not clear in
the slightest as to how this is supposed to work.

(Only tested with a USB 1.0 audio device which does produce
audio with this but in a very "cutting" pattern; but perhaps
that's just what multi_audio_test does?)