Update patch set 3
Patch Set 3:

> Patch Set 2:
> > Well, yes, submitting more than 64 blocks can be a bad idea, if you actually have to wait for all of them. So I agree with you here. Still, you could do several commits of 64 blocks at once.
> But that risks running into the same issue where a new transaction cannot be queued because it is waiting for a block to be written back. The more you enqueue at once, the more you risk running into this issue.

I might still be on vacation mentally, but I don't see how this could happen. As soon as a transaction is finished (ie. marked to be written back in BFS), new transactions touching the same blocks must always wait for them to be actually written back, or else you wouldn't have a working journal.

BFS uses sub transactions to reduce the time needed to wait for blocks being written back, but that should be independent from the actual block writer.

> So even this point alone is enough to justify not submitting everything at once IMO. At any rate, this new logic is quite literally 200x more performant, so if we still need more, we can return to it yet again :)

I think it would be best to perform actual benchmarks with variants of this code, to know for sure. This stuff is complex enough that you can expect having forgotten about something (unless you're very into it, which I'm definitely not anymore).

Patch-set: 3