BLocker: Do not even create a semaphore in "benaphore" mode.
Instead, we now just use a pthread_mutex, which will be
entirely atomic and not touch the kernel at all if there
is no lock contention, and will also not hit the kernel
for BLocker creation, so this is potentally a pretty
large performance improvement.

This also has the advantage of not using up a significant
number of the system semaphores to service BLockers, which
might help with some of Web+'s semaphore greed...

When we are using benaphore-style, the semaphore is still
acquired/released before/after each lock/unlock, thus
preserving compatibility with older applications. However,
applications that expected the semaphore to be somehow
useful in "benaphore" mode (and I don't know how they
would have, behavior would likely have already broken
vs. BeOS) will be broken by this change; so we need to
consider that before merging in some way.

The other potential side-effect of this change is that
anyone calling Sem() as a way to check if the BLocker is
alive will now fail incorrectly. BeOS did not have an
InitCheck() method, but their header has a comment noting
that the use of Sem() is /* -- For debugging (only!) -- */,
so do we need to worry about that too much? (Perhaps
we should rename Sem() if not R5_COMPATIBLE and fix
the broken applications?)