Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > I might still be on vacation mentally, but I don't see how this could happen. As soon as a transaction is finished (ie. marked to be written back in BFS), new transactions touching the same blocks must always wait for them to be actually written back, or else you wouldn't have a working journal.
> 
> I know; that's what I'm referring to here: the block_cache has the ability to write back individual blocks that were part of a previous transaction, in a new transaction: https://github.com/haiku/haiku/blob/master/src/system/kernel/cache/block_cache.cpp#L1054

No, you misread that one. Those are the blocks of the previous transaction, not a sub transaction: once a transaction is done, the still pending blocks of an already ended earlier transaction have to be written back. It will always wait for those, and will therefore benefit if all of those blocks are written back as fast as possible.

So I was incorrect in my statement above, too: individual blocks don't cause waiting for an earlier transaction.

> > I think it would be best to perform actual benchmarks with variants of this code, to know for sure. This stuff is complex enough that you can expect having forgotten about something (unless you're very into it, which I'm definitely not anymore).
> 
> I did -- before this commit, we had the 2-second hardcoded timeout; after this commit, heavy disk I/O will often see 200-300ms timeouts on an HDD, or 10-30ms timeouts on an SSD, so a 10x-200x improvement. Those numbers weren't theoretical :)

I'm not talking about the timeouts, those are indeed an obvious improvement. But while you're at it, why not squeeze as much out of it as possible? It's still low hanging fruit, after all :-)

Patch-set: 3