HaikuDepot: Remove User Creation
Because the user-creation process has become more
complex, the person using HaikuDepot is no longer
able to create a new user from the HaikuDepot C++
desktop application; instead they are directed to
the web user interface and they should create the
user there.  The APIs are still in place to
support creation of users in the desktop software
later if/when there is enough input to further
develop and maintain it.  At the same time, the
login process has been improved a little with a
spinner, better error handling etc...