Update patch set 1
Patch Set 1:

Yes, after looking into the AHCI spec there is nothing to change here. I'm not here just to please the analyzer, so let's just leave it as it is.

A way to fix it would be to declare cfis as an union between uint8_t[0x40] and various structures to be used for the different command types, but we don't have such structures declared anywhere (and with bitfields they would be a little annoying to do anyway). Declaring an union between uint8_t[0x40] and uint8_t[0x16] is not great, so...

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>