Update patch set 1
Patch Set 1:

(3 comments)

For WebKit, as far as I'm aware we have an issue with running out of ports, but not running out of semaphores. In any case, using a lighter synchronization primitive here is definitely a good idea, however I don't find the implementation that great.

As mentionned in inline comments, we can make it a lot cleaner if we add a custom type of pthread_mutex directly implementing the behavior we need here, then implementing BLocker on top of it will be more trivial and easier to maintain since all the locking logic will be all in one place. The code as it is now is a bug waiting to pop whenever someone changes the behavior of pthreads without remembering to look here.

How comes our pthread internals are exposed to other headers in this way, anyway? They should rather be an opaque structure?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>