Update patch set 1
Patch Set 1: Code-Review-1

Since the point of the benaphore is not to hit the kernel when there is no contention, the performance gain should be marginal in operation in this case -- only creation and deletion should be substantially faster (blocking/unblock is also faster, but shouldn't happen that often).

IMO the performance and resource gains are not worth applying this change for now -- there aren't that many BLockers around that this is really hurtful, and the compatibility issues could be potentially harmful.

I would prefer if we make this change when we do a major API overhaul, probably for R2.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1