Menu tracking: Re-apply Stefano's patch from #1621
The idea is to simplify menu tracking by using Mouse hooks.

I had to manually apply the patch and I know a lot has changed since it was
written.

This compiles but functionality is totally broken. One menu will open and track
mouse movements, but no other menus open, selecting items doesn't trigger
anything and the one menu won't close. Obviously a ton of needed code has been
removed or commented out. It will take me a while to sort it out, but I at
least wanted to give it a try and get it pushed up to Gerrit.