kernel/vm: Make the page_scrubber wait on a semaphore.
Since it now will block if there is nothing to do, there's no
real reason to leave the timeout at 100ms, so decrease it to 50.
Additionally bump the SCRUB_SIZE to 32 pages, because why not.
This effectively quadruples the rate at which the page scrubber
can scrub pages.

Following this change, the page scrubber seems to run on a
"mostly idle" system every ~250 ms on average. So this seems
more than worth it.