Terminal: Allow use of Option as Meta key
Add a configuration setting that allows the left Option key to be used as a
Meta key, and add support for the Escape sequences that control the Meta key's
behaviour.

TermWindow now maintains a copy, shared by all its component TermViews, of the
current keymap, and will automatically reload the keymap when it appears to
have changed (by monitoring the user's keymap file).

Fixes #15294.

Change-Id: I150b640b7b18384d56ab2fb017bf16ce8bdbdd78