Media Kit: Clear the BufferCache of buffers for disconnected clients
Without this, the BufferCache keeps a reference to these buffers inside the
media_addon_server until the media_addon_server quits, which is pretty much
never.

This was the bug no one thought could be fixed.

Should fix #4954 and #14755, and possibly #13614 and #14047, though I think
they may be something else.