Update patch set 1
Patch Set 1:

It's deviating from BeOS behavior, but in a good way. There is no good reason I can think of that this should not always go through BView::MessageReceived(). It potentially breaks existing applications, though -- while you should always call the super class in MessageReceived(), the hooks were working fine without before.

I disagree with Simon's assessment, that you should not get the message from the hook to get to all of its fields, but that doesn't really matter (I'm not a fan of the API either, though).

The window needs to know the inner workings of the view, and there are a lot more things that windows call in the views, like drawing and pulse messages.

However, I don't see any actual problem that this solves. It just changes behavior, and does make a bit more sense, but it basically comes down to how it's documented, and how it used to work; I would say either way is okay. It doesn't create a situation that you couldn't have possibly solved before.

Because of this, I would lean towards keeping the existing behavior.

Patch-set: 1
CC: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>