Update patch set 1
Patch Set 1:

It can break things in the other direction, too. Imagine an application deciding to manually forward the B_KEY_DOWN message to some views in the window as it is handling keyboard events in its own quirky way. It could do that and very well not expect BKeyDown to be called (after some testing on BeOS and finding that the implementation is not in line with the docs).

So, if this allows getting something done, I'd be ok with it, but if not, we should at least be ready to revert it if we find out it does indeed break something. This way, at least people at haikuports and other 3rd party developers will not be too annoyed. And we can reconsider merging it later once all issues have been solved.

Patch-set: 1