Update patch set 1
Patch Set 1:

Can't remember why I removed (or deprecated) the deselect hooks but I have a feeling I did it for a reason -- I think perhaps to remove the possibility of the driver attempting to do a notify while it was being deselected and causing an inversion in locking order. -- That's also why I made the drivers return the currently active events from the select hook instead of notifying.

Patch-set: 1
CC: Gerrit User 1000113 <1000113@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>