HaikuDepot: Improvements to User Ratings
Change to match the web user interface behaviour;
the user ratings for all architectures are shown
and the list of user ratings is restricted to the
actual repository in use.  This is important
where two repositories have the same package with different versions.  Also the architecture
is displayed with the version in the package info
page.