Update patch set 1
Patch Set 1:

Yes, there would be a lot of other places to review. But I'm not sure this is the right level to handle this, so I didn't want to make all the changes and then get told that another way is preferred.

But doing it at an upper level is less easy, because this method may already send a reply to the app_server link in some cases, and we must not send a second one (the app_server link does not like to be desynchronized). So we need a try/catch in each case of the dispatch, or refactor the whole error handling here to make sure only one error reply can be written in the app_server link.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>