Mail: Better placement of signature menu
When clicking the "Signature" button in the toolbar, no mouse
coordinates are available. The current hack is to open the menu
in the center of the tool bar, which is normally no way near
where the user has clicked the button. Not nice.

Change that to find the button coordinates and place the menu a bit
to the right of its center.

Fall back to the old hack, if the signature button wasn't found for
some reason...