ext2: add test script for ext2 filesystem.
based on a test script from FreeBSD.
this uses fsx and fstorture to exercise the fs, mkfs.ext3 to initialize,
e2fsck to check the integrity. single-threaded test is OK, multi-threaded is not.
the following blocksizes and features are tested: 1K, 2K and 4K; huge_file,
dir_nlink, extent, 64bit, metadata_csum.