Update patch set 1
Patch Set 1:

> Let's just work on beta2 and patch the single broken application. If there were more than one, maybe we would need to do something in the OS itself, but there are not.

That's not the right way to do things. We should aim for the same level of backwards compatibility as Windows, so even the odd single app should be handled gracefully. Take it as training for when we have thousands of apps and need to do similar changes again.

This is not a "move fast and break things" project. Especially when we have a working solution here.

Let's indeed work on beta2 so we can bump our API version then, and then the change becomes effective for all newly built apps and libraries.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>