Update patch set 1
Patch Set 1: Code-Review+2

Should we add a BControlLook::GetTintForBase() or something that automatically computes an appropriate tint?
If we are going to do this in several places, I fear that we may find out cases where it doesn't work and we want to change the condition again, it would be great if we could do it in one single place in BControlLook next time.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2