Update patch set 1
Patch Set 1:

> That's not the right way to do things. We should aim for the same level of backwards compatibility as Windows, so even the odd single app should be handled gracefully. Take it as training for when we have thousands of apps and need to do similar changes again.

I don't think we need "training" here. As this patch shows, we already have a way to gracefully handle such changes if needed. I don't think it is needed here.

Windows also has begun breaking compatibility with Windows 2000 apps and the like. Even they decided there are limits.

> This is not a "move fast and break things" project. Especially when we have a working solution here.

I think we need to be able to break things once and a while to improve other things. We already broke FS API since BeOS (userlandfs allows source compatibility to be preserved, but not binary compatibility), we broke driver API for SMAP and other such changes, etc. etc.

The fact that any process on the system could clone any area at all was a massive security hole. A single flag to allow it still is not great, but at least is way better than we had before.

Once again, if there was more than one app, I'd agree we'd need to change something. But the app is open source and easily changed.

If it was possible to build beta1 still, I'd push a change there to allow the flag to at least be set (unused though) so we can fix it right now. But of course beta1 can't be built because the Buildbot is still broken here, which is also a reason beta2 is stalled.

> Let's indeed work on beta2 so we can bump our API version then, and then the change becomes effective for all newly built apps and libraries.

But as shown, 99.99% of apps and libs can use this protection. Making the whole system less secure for the sake of one app is a bad idea IMO.

Patch-set: 1
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>