Update patch set 2
Patch Set 2:

> But as shown, 99.99% of apps and libs can use this protection. Making the whole system less secure for the sake of one app is a bad idea IMO.

Indeed, but this is just qbout making it unsecure until the next API/ABI bump, right? So if that's really a concern for you, we should just bump the API sooner, and new apps benefit from it. We are in beta, so I would say we are not in an emergency situation.

And yes, Microsoft is starting to drop compatibility with Windows 2000, which was released about the same time as BeOS R5. This is not at all the same as dropping support for apps built a few month ago and which may already be too new to run on beta1.

Patch-set: 2