Update patch set 2
Patch Set 2:

> Indeed, but this is just qbout making it unsecure until the next API/ABI bump, right? So if that's really a concern for you, we should just bump the API sooner, and new apps benefit from it. We are in beta, so I would say we are not in an emergency situation.

A new ABI bump *and* a rebuild of all packages; something that we have not done for quite a while, and would probably take a few weeks with our current build setup...

> And yes, Microsoft is starting to drop compatibility with Windows 2000, which was released about the same time as BeOS R5. This is not at all the same as dropping support for apps built a few month ago and which may already be too new to run on beta1.

There are not many (if any?) of said apps. HaikuPorts is still built against beta1, and of course all users running on beta1 will not encounter any breakage, because the CLONEABLE_AREA changes are not there; and of course we will fix said app for beta2. And, again, this is still *one* application, not apps, plural.

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>