Update patch set 2
Patch Set 2:

> A new ABI bump *and* a rebuild of all packages; something that we have not done for quite a while, and would probably take a few weeks with our current build setup...

Then we need to scale up the infra. See what I mean about "we need training about these things"?

> > And yes, Microsoft is starting to drop compatibility with Windows 2000, which was released about the same time as BeOS R5. This is not at all the same as dropping support for apps built a few month ago and which may already be too new to run on beta1.
> 
> There are not many (if any?) of said apps. HaikuPorts is still built against beta1, and of course all users running on beta1 will not encounter any breakage, because the CLONEABLE_AREA changes are not there; and of course we will fix said app for beta2. And, again, this is still *one* application, not apps, plural.

So it is currently not even possible to build apps that use the new flag with our setup, and we're already deprecating everything that is built on our current packaging infra. I don't care if this time it's one single app, this is something we should never do.

Let's do the things in order, and release a version with the new API before we start deprecating the previous one. Yes, it's slower and in the meantime apps will be insecure, but in the meantime people can continue using their system.

In your recent changes, for various reasons, many apps have been broken:
- cmake would not work (reverted yesterday)
- Vision cannot connect to SSL servers with untrusted certificates and offer no way to override that (in contrary to your comment in the commit that "all apps have been adjusted already")
- This change breaks yet another app

These things add up quickly to making it annoying to use Haiku and dangerous to upgrade the system. Even if each time it's "just one app". So please consider migration paths and how to make these things happen more smoothly. Send warnings to haiku-3rdparty about upcoming changes. Add __attribute__((deprecated)) to functions that are going away, or make them print warning to the console when they catch something (just like we had the "resizing mode and flags swapped" in BView constructor for as far as I remember). Bump the API if needed when a change is breaking things.

Let's get all this process set up even this time where it's just one app. So it's ready to roll when we make a major change that affects everyone.

Patch-set: 2