Update patch set 2
Patch Set 2:

> Patch Set 2:
> > Then we need to scale up the infra. See what I mean about "we need training about these things"?
> 
> Yes, but more infra costs more money, and time to maintain (the buildmaster is a pain to maintain as it is, etc.)

How exactly is the buildmaster a pain to maintain? Adding a new builder requires exactly two files, one holding the connection and setup details and a private key. Targetting a different Haiku version means just swapping out the initial packages.

I have not needed to touch this setup since forever, it just runs. Since noone even bothered to inform me about the server change and builds continue to run, I presume that it didn't need any manual work either. Except for the virtio hangs, which required resetting the builder VMs from time to time (and is now fixed), these require exactly zero maintenance.

So yes, more infra costs more money and that is a real concern. But please don't hide an unwillingness to do proper changes behind some imaginary technical burden.

Since I've been sponsoring the builders at my expense since setting them up, it would be the first cost for builders Haiku would actually need to pay. If it is about temporarily speeding things up, such builders could just as well be added temporarily for a full rebuild and then removed.

> > > > And yes, Microsoft is starting to drop compatibility with Windows 2000, which was released about the same time as BeOS R5. This is not at all the same as dropping support for apps built a few month ago and which may already be too new to run on beta1.
> 
> Again, the buildmaster is still targeting beta1. All packages currently installable from HaikuPorts will build and run on beta1.

So either spin up a new buildmaster (it's trivial, it's a container afterall), or retarget the existing one once the next release is done. Doing a transition release is how these things are usually done. I know it takes longer, but what's the hurry?

> > I don't care if this time it's one single app, this is something we should never do.
> 
> Look, the #3 complaint I hear about Haiku after "it doesn't run on my hardware" and "the web browser doesn't load X website properly" is "it isn't secure / has no multiuser segmentation / etc." We need to address these concerns to get more adoption, and thus more momentum, funding, etc. etc. etc.

This really is a question of principle. Do you take shortcuts or not? For me as a developer possibly targetting a platform, I want to know such things. If I realize that the people maintaining that platform will try hard to do changes properly, so that my software is not likely to break arbitrarily, this makes the platform attractive. If I have to fear that it's a moving target that someone will simply change underneath me, this is a big concern and possible reason not to invest in the platform at all. This then leads to the burden of porting things remaining in people at HaikuPorts and never spreading out to the actual projects. This IMHO is far worse for the momentum.

And really, as long as everything runs as root and we even have the kernel debugger scriptable and open, the security argument is somewhat moot. That doesn't mean we shouldn't invest in it, and I agree with the clone area change in general, but there definitely is no real urgency as the more fundamental changes will take more time to settle anyway.

> So to add a patch like this that essentially negates the benefits of such a change (since e.g. WebPositive depends on WebKit, which comes from HaikuPorts and so will have the lower API version, and so not get the protection, etc.) 

It doesn't negate the benefit, it postpones it for a proper introduction with a proper process. And we should absolutely strive for that IMO as it sends a signal: We do care and things will not just break.

The whole point of "it's only one app" is a strange attitude IMO. You can't possibly know that. Someone may be using Haiku with some software you don't happen to know and that software may be important to that person. Or someone may be in the process of developing something right now and would be put off be such a breakage. You simply cannot really forsee the entirety of it, and that's what demands a proper introduction process for such changes.

> > Let's do the things in order, and release a version with the new API before we start deprecating the previous one. Yes, it's slower and in the meantime apps will be insecure, but in the meantime people can continue using their system.

Yes, please let's do things in order.

> People can continue using their system ... on the beta that we set aside for "maybe better for production use than nightly builds."

That is not the point. For it to be proper, there would need to be a new release with the change and the proper backwards compatibility in place, and only in the next release (at the earliest) that compat could be removed.

> Well, that one is just a straight-up bug in my change, it seems, unlike this clone-areas change where the breakage is due to an actual design choice in my change. So reverting is indeed the right way to go before I can figure out what was wrong with it.

This is the proper distinction of nigthly builds vs. release. In the nightly things may break and it is ok. But they break because of bugs, not because of API changes that didn't roll out properly.

> This area change was the only one that actually broke an application which was not previously broken, intentionally.

And that's all this argument is about. Intentional breakage should not just be dropped like this, it should be properly rolled out with a transition period. No big platform just breaks things like that, even when the affected audience is small.

It is different for changes that break apps that were really already broken but happened to still work by accident. But changing longstanding public API should simply not be done without a proper release process.

Patch-set: 2
CC: Gerrit User 1000017 <1000017@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>